_newcode refactored. Just a few more tweaks to go.
